- fixed some formatting issues - updated names of classes - everything compiles now

+++ src/jgrl/struct/GeoTuple_F32.java
@@ -73,7 +73,7 @@ public abstract class GeoTuple_F32<T extends GeoTuple_F32> extends GeoTuple<T> {
 	 * @return norm.
 	 */
 	public float norm() {
-		return (float) Math.sqrt(normSq());
+		return (float) (float) Math.sqrt(normSq());
 	}
 
 	/**
